fix: creating payloads from retrieved items in integrity check
