+++ packages/snjs/lib/services/Sync/Account/ResponseResolver.ts
@@ -4,11 +4,11 @@ import {
   ImmutablePayloadCollection,
   ImmutablePayloadCollectionSet,
   CopyPayload,
+  filterDisallowedRemotePayloads,
 } from '@standardnotes/payloads'
 import { SyncResponse } from '@Lib/services/Sync/Response'
 import { DeltaClassForSource } from '@Payloads/deltas/generator'
 import { HistoryMap } from '@Lib/services/History/HistoryMap'
-import { filterDisallowedRemotePayloads } from '../Filter'
 
 /**
  * Given a remote sync response, the resolver applies the incoming changes on top
